Additional refactoring.
